+++ AssemblyInfo.cs
@@ -6,8 +6,8 @@ using System.Windows;
 [assembly: AssemblyCompany("Phạm Duy")]
 [assembly: AssemblyProduct("SmartMacroAI")]
 [assembly: AssemblyCopyright("Created by Phạm Duy - Giải pháp tự động hóa thông minh.")]
-[assembly: AssemblyVersion("1.6.0.0")]
-[assembly: AssemblyFileVersion("1.6.0.0")]
+[assembly: AssemblyVersion("1.6.2.0")]
+[assembly: AssemblyFileVersion("1.6.2.0")]
 
 [assembly: ThemeInfo(
     ResourceDictionaryLocation.None,